requirements.txt
streamlit
google-generativeai==0.3.2
python-dotenv

torch>=2.0

diffusers==0.28.0
transformers==4.38.2
huggingface-hub==0.21.4

accelerate
safetensors
Pillow
protobuf==3.20.3
